pyproject.toml
[project]
name = "omnicloud-ops-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.18",
    "claude-agent-sdk>=0.1.18",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "python-telegram-bot>=21.0",
]
